requirements.txt
requests>=2.31.0
pandas>=2.1.0
pyarrow>=14.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
pytest>=7.4.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
scikit-learn>=1.3.0
joblib>=1.3.0
